pyproject.toml
[project]
name = "pynspd"
dynamic = ["version"]
description = "Python-библиотека для работы с https://nspd.gov.ru (ex-ПКК)"
readme = "docs/README.md"
license = {file = "LICENSE"}
authors = [
    { name = "Aleksandr Litovchenko", email = "yazmolod@gmail.com" }
]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
requires-python = ">=3.10"
dependencies = [
    "geojson-pydantic>=1.2.0",
    "hishel>=0.1.1",
    "httpx>=0.28.1",
    "mercantile>=1.2.1",
    "numpy>1.24.4",
    "pydantic>=2.10.4",
    "pyproj>=3.7",
    "shapely>=2.0.6",
]

[project.urls]
Homepage = "https://yazmolod.github.io/pynspd/"
Repository = "https://github.com/yazmolod/pynspd"
Issues = "https://github.com/yazmolod/pynspd/issues"
Changelog = "https://github.com/yazmolod/pynspd/releases"

[project.optional-dependencies]
cli = [
    "geopandas>=1.0.1",
    "openpyxl>=3.1.5",
    "typer>=0.15.2",
]
socks = [
    "socksio>=1.0.0",
]
redis = [
    "redis==6.0.0",
]
sqlite = [
    "anysqlite>=0.0.5",
]

[project.scripts]
pynspd = "pynspd.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "folium>=0.19.4",
    "geopandas>=1.0.1",
    "pandas>=2.2.3",
    "ipykernel>=6.29.5",
    "jinja2>=3.1.5",
    "pre-commit>=4.0.1",
    "tqdm>=4.67.1",
    "hatch>=1.14.0",
]
tests = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.0",
    "coverage>=7.6.10",
    "coveralls[yaml]>=4.0.1",
    "pytest-httpx>=0.35.0",
    "python-dotenv>=1.1.0",
]
lint = [
    "ruff>=0.8.4",
    "mypy>=1.14.0",
]
docs = [
    "mkdocs-material>=9.6.2",
    "mkdocstrings[python]>=0.28.1",
]

[tool.mypy]
enable_incomplete_feature = "NewGenericSyntax"

[tool.pytest]
asyncio_mode = "auto"

[tool.hatch.version]
path = "src/pynspd/__init__.py"
